+++ lib/src/impl_ffi/impl_ffi.utils.dart
@@ -25,7 +25,10 @@ class _EvpPKey implements ffi.Finalizable {
   static const _externalSizeEstimate = 4096;
 
   static final _finalizer = ffi.NativeFinalizer(
-    ssl.addresses.EVP_PKEY_free.cast(),
+    ffi.Native.addressOf<
+          ffi.NativeFunction<ffi.Void Function(ffi.Pointer<EVP_PKEY>)>
+        >(ssl.EVP_PKEY_free)
+        .cast(),
   );
 
   final ffi.Pointer<EVP_PKEY> _pkey;
@@ -308,12 +311,14 @@ extension on _Scope {
 
   ffi.Pointer<CBS> createCBS(List<int> data) {
     final cbs = this<CBS>();
-    ssl.CBS_init(cbs, dataAsPointer(data), data.length);
+    cbs.ref.data = dataAsPointer(data);
+    cbs.ref.len = data.length;
     return cbs;
   }
 
   ffi.Pointer<CBB> createCBB([int sizeHint = 4096]) {
-    final cbb = this<CBB>();
+    final cbbSize = nativeWebcryptoGetCbbSize();
+    final cbb = allocate<ffi.Uint8>(cbbSize).cast<CBB>();
     ssl.CBB_zero(cbb);
     _checkOp(ssl.CBB_init(cbb, sizeHint) == 1, fallback: 'allocation failure');
     defer(() => ssl.CBB_cleanup(cbb));