+++ lib/src/impl_js/impl_js.utils.dart
@@ -131,11 +131,15 @@ Future<T> _handleDomException<T>(
 }) async {
   try {
     return await fn();
-  } on subtle.JSDomException catch (e) {
-    throw _translateDomException(
-      e,
-      invalidAccessErrorIsArgumentError: invalidAccessErrorIsArgumentError,
-    );
+  } catch (e) {
+    // ignore: invalid_runtime_check_with_js_interop_types
+    if (e is JSAny && e.isA<subtle.JSDomException>()) {
+      throw _translateDomException(
+        e as subtle.JSDomException,
+        invalidAccessErrorIsArgumentError: invalidAccessErrorIsArgumentError,
+      );
+    }
+    rethrow;
   }
 }