+++ lib/src/testing/testing.dart
@@ -30,6 +30,8 @@ import 'webcrypto/rsassapkcs1v15.dart' as rsassapkcs1v15;
 // Other test files, that don't use TestRunner
 import 'webcrypto/random.dart' as random;
 import 'webcrypto/digest.dart' as digest;
+import 'regression/derive_bits_zero_length.dart' as derive_bits_zero_length;
+import 'regression/issue_60_trailing_bytes.dart' as issue_60_trailing_bytes;
 
 /// Test runners from all test files except `digest.dart` and
 /// `random.dart`, which do not use [TestRunner].
@@ -58,6 +60,8 @@ void runAllTests(
     for (final r in _testRunners) ...r.tests(),
     ...random.tests(),
     ...digest.tests(),
+    ...issue_60_trailing_bytes.tests(),
+    ...derive_bits_zero_length.tests(),
   ];
 
   for (final (:name, :test) in allTests) {