+++ lib/src/webcrypto/webcrypto.ecdh.dart
@@ -92,7 +92,7 @@ final class EcdhPrivateKey {
   /// }
   /// ```
   ///
-  /// [1]: https://datatracker.ietf.org/doc/html/rfc5208
+  /// [1]: https://www.rfc-editor.org/rfc/rfc5208
   static Future<EcdhPrivateKey> importPkcs8Key(
     List<int> keyData,
     EllipticCurve curve,
@@ -155,7 +155,7 @@ final class EcdhPrivateKey {
   /// }
   /// ```
   ///
-  /// [1]: https://www.rfc-editor.org/rfc/rfc7518.html#section-6.2
+  /// [1]: https://www.rfc-editor.org/rfc/rfc7518#section-6.2
   static Future<EcdhPrivateKey> importJsonWebKey(
     Map<String, dynamic> jwk,
     EllipticCurve curve,
@@ -246,7 +246,7 @@ final class EcdhPrivateKey {
   ///   print(PemCodec(PemLabel.privateKey).encode(exportedPkcs8Key));
   /// }
   /// ```
-  /// [1]: https://datatracker.ietf.org/doc/html/rfc5208
+  /// [1]: https://www.rfc-editor.org/rfc/rfc5208
   Future<Uint8List> exportPkcs8Key() => _impl.exportPkcs8Key();
 
   /// Export the [EcdhPrivateKey] as a [JSON Web Key][1].
@@ -270,7 +270,7 @@ final class EcdhPrivateKey {
   ///   print(jsonEncode(exportedPrivateKey));
   /// }
   /// ```
-  /// [1]: https://www.rfc-editor.org/rfc/rfc7518.html#section-6.2
+  /// [1]: https://www.rfc-editor.org/rfc/rfc7518#section-6.2
   Future<Map<String, dynamic>> exportJsonWebKey() => _impl.exportJsonWebKey();
 }
 
@@ -364,7 +364,7 @@ final class EcdhPublicKey {
   /// }
   /// ```
   ///
-  /// [1]: https://www.rfc-editor.org/rfc/rfc7518.html#section-6.2
+  /// [1]: https://www.rfc-editor.org/rfc/rfc7518#section-6.2
   static Future<EcdhPublicKey> importJsonWebKey(
     Map<String, dynamic> jwk,
     EllipticCurve curve,
@@ -409,6 +409,6 @@ final class EcdhPublicKey {
   ///   print(jsonEncode(exportedPublicKey));
   /// }
   /// ```
-  /// [1]: https://www.rfc-editor.org/rfc/rfc7518.html#section-6.2
+  /// [1]: https://www.rfc-editor.org/rfc/rfc7518#section-6.2
   Future<Map<String, dynamic>> exportJsonWebKey() => _impl.exportJsonWebKey();
 }